test: move model regressions out of generated files
